backend/requirements.txt
cachetools==3.1.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-SocketIO==4.2.1
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-core==1.0.3
google-cloud-datastore==1.10.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
protobuf==3.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
Werkzeug==0.16.0
gunicorn==19.9.0
python-jose==3.1.0